requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.3  # Replaces langchain-openai
langgraph>=0.0.26
langchain-community>=0.0.10
pydantic>=2.5.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0    # Replaces openai
tenacity>=8.2.2